ai-governance-platform-full-enhanced/backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
requests==2.31.0
python-dotenv==1.0.0
pydantic==1.10.11
sqlmodel==0.0.7
python-multipart==0.0.6
pillow==10.0.0
pytesseract==0.3.10
pdf2image==1.16.3
python-magic==0.5.0
aiohttp==3.8.5
sentence-transformers==2.2.2
torch>=1.13.1
